refactor(amount): Changed amount type
Changed amount type to BigDecimal to avoid gaps with Double

+++ src/main/java/com/kairos/technicalproof/domain/core/Amount.java
@@ -3,9 +3,11 @@ package com.kairos.technicalproof.domain.core;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
+import java.math.BigDecimal;
+
 @RequiredArgsConstructor
 @Getter
 public class Amount {
-    private final Double value;
+    private final BigDecimal value;
     private final String currency;
 }
