+++ src/main/java/com/kairos/technicalproof/infrastructure/primary/api/model/price/FindPriceResponse.java
@@ -3,6 +3,7 @@ package com.kairos.technicalproof.infrastructure.primary.api.model.price;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 @RequiredArgsConstructor
@@ -13,5 +14,5 @@ public class FindPriceResponse {
     private final Long priceListId;
     private final Instant initialDate;
     private final Instant finalDate;
-    private final Double amount;
+    private final BigDecimal amount;
 }
